+++ react/OrderCoupon.tsx
@@ -1,95 +1,105 @@
-import React, { createContext, ReactNode, useContext, useState } from 'react'
+import React, {
+  createContext,
+  ReactNode,
+  useContext,
+  useState,
+  useCallback,
+  useRef,
+  useEffect,
+} from 'react'
 import { compose, graphql } from 'react-apollo'
 import { useOrderQueue } from 'vtex.order-manager/OrderQueue'
+import { useOrderForm } from 'vtex.order-manager/OrderForm'
 
 import InsertCoupon from './graphql/insertCoupon.graphql'
 
 interface Context {
-  insertCoupon: (coupon: string) => PromiseLike<void>
   coupon: string
-  isShowingPromoButton: boolean
+  setCoupon: (coupon: string) => void
+  insertCoupon: (coupon: string) => void
   errorKey: string
-  changeCoupon: (coupon: string) => void
-  handleCouponChange: (evt: any) => void
-  resetCouponInput: () => void
-  setIsShowingPromoButton: (value: boolean) => void
-  submitCoupon: (evt: any) => void
+  setErrorKey: (errorKey: string) => void
+  showPromoButton: boolean
+  setShowPromoButton: (value: boolean) => void
 }
 
 interface OrderItemsProviderProps {
   children: ReactNode
   InsertCoupon: any
+  CouponQuery: any
   coupon: string
   changeCoupon: (coupon: string) => void
 }
 
-const NO_ERROR = ''
-const CODE_DOESNT_EXIST = `CodeDoesntExist`
+const couponKey = 'coupon'
+const noError = ''
 
 const OrderCouponContext = createContext<Context | undefined>(undefined)
 
 export const OrderCouponProvider = compose(
   graphql(InsertCoupon, { name: 'InsertCoupon' })
 )(({ children, InsertCoupon }: OrderItemsProviderProps) => {
-  const { enqueue } = useOrderQueue()
-  const [coupon, setCoupon] = useState('')
-  const [isShowingPromoButton, setIsShowingPromoButton] = useState(true)
-  const [errorKey, setErrorKey] = useState(NO_ERROR)
+  const { enqueue, listen } = useOrderQueue()
+  const { orderForm, setOrderForm } = useOrderForm()
+  const [coupon, setCoupon] = useState(orderForm.marketingData.coupon || '')
+  const [showPromoButton, setShowPromoButton] = useState(true)
+  const [errorKey, setErrorKey] = useState(noError)
 
-  const changeCoupon = (coupon: string) => {
-    setCoupon(coupon)
-  }
+  const isQueueBusy = useRef(false)
+  useEffect(() => {
+    const unlisten = listen('Pending', () => (isQueueBusy.current = true))
+    return unlisten
+  })
+  useEffect(() => {
+    const unlisten = listen('Fulfilled', () => (isQueueBusy.current = false))
+    return unlisten
+  })
 
-  const handleCouponChange = (evt: any) => {
-    evt.preventDefault()
-    const newCoupon = evt.target.value.trim()
-    changeCoupon(newCoupon)
-  }
+  const insertCoupon = useCallback(
+    (coupon: string) => {
+      const marketingData = { ...orderForm.marketingData, coupon }
 
-  const resetCouponInput = () => {
-    changeCoupon('')
-    setIsShowingPromoButton(false)
-  }
+      setOrderForm({ marketingData })
 
-  const submitCoupon = (evt: any) => {
-    evt.preventDefault()
-    setErrorKey(NO_ERROR)
-    insertCoupon(coupon)
-  }
+      setShowPromoButton(true)
 
-  const insertCoupon = (coupon: string) => {
-    return enqueue(async () => {
-      const mutationResult = await InsertCoupon({
-        variables: {
-          text: coupon,
-        },
-      })
+      const task = async () => {
+        const {
+          data: { insertCoupon: newOrderForm },
+        } = await InsertCoupon({
+          variables: {
+            text: coupon,
+          },
+        })
 
-      const newCoupon = mutationResult.data.insertCoupon.code
-        ? mutationResult.data.insertCoupon.code
-        : ''
-
-      if (newCoupon) {
-        setCoupon(newCoupon)
-        setIsShowingPromoButton(true)
-      } else {
-        setErrorKey(CODE_DOESNT_EXIST)
+        return newOrderForm
       }
-    })
-  }
+
+      enqueue(task, couponKey)
+        .then((newOrderForm: OrderForm) => {
+          if (!isQueueBusy.current) {
+            setOrderForm(newOrderForm)
+          }
+        })
+        .catch((error: any) => {
+          if (!error || error.code !== 'TASK_CANCELLED') {
+            throw error
+          }
+        })
+    },
+    [InsertCoupon, enqueue, orderForm, setOrderForm]
+  )
 
   return (
     <OrderCouponContext.Provider
       value={{
-        insertCoupon: insertCoupon,
         coupon: coupon,
-        changeCoupon: changeCoupon,
-        handleCouponChange: handleCouponChange,
-        resetCouponInput: resetCouponInput,
-        setIsShowingPromoButton: setIsShowingPromoButton,
-        isShowingPromoButton: isShowingPromoButton,
-        submitCoupon: submitCoupon,
+        setCoupon: setCoupon,
+        insertCoupon: insertCoupon,
+        showPromoButton: showPromoButton,
+        setShowPromoButton: setShowPromoButton,
         errorKey: errorKey,
+        setErrorKey: setErrorKey,
       }}
     >
       {children}