+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.2.2] - 2019-08-29
 
+## Added
+
+- Use OrderQueue and OrderForm
+
+## Changed
+
+- Move handleCouponChange, resetCouponInput and submitCoupon to UI component
+- Rename flag isShowingPromoButton to showPromoButton
+
 ## [0.2.1] - 2019-08-27
 
 ## [0.2.0] - 2019-08-26